feat: Refactorisation complète des pages protégées vers Server Components
- Migration de toutes les pages utilisant useAuth vers Server Components
- Élimination des race conditions d'authentification côté client
- Séparation claire entre Server Components (auth) et Client Components (UI)

## Pages refactorisées:
- ✅ dashboard/page.tsx → Server Component + dashboard-page-client.tsx
- ✅ dashboard/followups/page.tsx → Server Component + followups-page-client.tsx
- ✅ admin/settings/followups/page.tsx → Server Component + followup-settings-page-client.tsx
- ✅ admin/mailboxes/[id]/page.tsx → Server Component + mailbox-detail-page-client.tsx

## Architecture finale:
- Server-side authentication avec redirect() sécurisé
- Vérification des rôles avant rendu des pages
- Performance améliorée avec SSR
- Plus de délais de chargement ou états loading bloquants

## Mise à jour documentation:
- Ajout du pattern Server + Client Component dans CLAUDE.md
- Guide d'architecture pour futures pages
