+++ app/(protected)/dashboard/page.tsx
@@ -1,224 +1,43 @@
-"use client";
+import { redirect } from "next/navigation";
+import { createClient } from "@/lib/supabase/server";
+import { DashboardPageClient } from "./dashboard-page-client";
 
-import { useAuth } from "@/lib/hooks/use-auth";
-import { Button } from "@/components/ui/button";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card";
-import { Badge } from "@/components/ui/badge";
-import { Mail, Users, BarChart3, Settings, LogOut, Plus } from "lucide-react";
+export default async function DashboardPage() {
+  // Server-side authentication check
+  const supabase = await createClient();
 
-export default function DashboardPage() {
-  const { user, signOut } = useAuth();
+  const {
+    data: { user },
+    error,
+  } = await supabase.auth.getUser();
 
-  const handleSignOut = async () => {
-    await signOut();
-  };
+  // Redirect if not authenticated
+  if (error || !user) {
+    redirect("/login");
+  }
 
-  const getRoleBadgeVariant = (role: string) => {
-    switch (role) {
-      case "administrateur":
-        return "destructive";
-      case "manager":
-        return "default";
-      case "utilisateur":
-        return "secondary";
-      default:
-        return "outline";
-    }
-  };
+  // Get user data from database
+  const { data: userData } = await supabase
+    .from("users")
+    .select("*")
+    .eq("id", user.id)
+    .single();
 
-  const getRoleLabel = (role: string) => {
-    switch (role) {
-      case "administrateur":
-        return "Administrateur";
-      case "manager":
-        return "Manager";
-      case "utilisateur":
-        return "Utilisateur";
-      default:
-        return role;
-    }
-  };
+  // Merge user data with proper typing
+  const userWithRole = userData
+    ? {
+        id: user.id,
+        email: user.email || userData.email,
+        full_name: userData.full_name,
+        role: userData.role,
+      }
+    : {
+        id: user.id,
+        email: user.email,
+        full_name: undefined,
+        role: undefined,
+      };
 
-  return (
-    <div className="min-h-screen bg-gray-50">
-      {/* Header */}
-      <header className="border-b bg-white shadow-sm">
-        <div className="mx-auto max-w-7xl px-4 sm:px-6 lg:px-8">
-          <div className="flex h-16 items-center justify-between">
-            <div className="flex items-center">
-              <Mail className="text-primary mr-3 h-8 w-8" />
-              <h1 className="text-xl font-semibold text-gray-900">
-                TrackedMail
-              </h1>
-            </div>
-            <div className="flex items-center space-x-4">
-              <div className="flex items-center space-x-2">
-                <span className="text-sm text-gray-700">
-                  {user?.full_name || user?.email}
-                </span>
-                <Badge variant={getRoleBadgeVariant(user?.role || "")}>
-                  {getRoleLabel(user?.role || "")}
-                </Badge>
-              </div>
-              <Button
-                variant="ghost"
-                size="sm"
-                onClick={handleSignOut}
-                className="text-gray-500 hover:text-gray-700"
-              >
-                <LogOut className="mr-2 h-4 w-4" />
-                Déconnexion
-              </Button>
-            </div>
-          </div>
-        </div>
-      </header>
-
-      {/* Main Content */}
-      <main className="mx-auto max-w-7xl py-6 sm:px-6 lg:px-8">
-        <div className="px-4 py-6 sm:px-0">
-          {/* Welcome Section */}
-          <div className="mb-8">
-            <h2 className="mb-2 text-2xl font-bold text-gray-900">
-              Tableau de bord
-            </h2>
-            <p className="text-gray-600">
-              Bienvenue {user?.full_name || user?.email}. Gérez vos e-mails
-              trackés et suivez vos performances.
-            </p>
-          </div>
-
-          {/* Stats Cards */}
-          <div className="mb-8 grid grid-cols-1 gap-6 md:grid-cols-2 lg:grid-cols-4">
-            <Card>
-              <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
-                <CardTitle className="text-sm font-medium">
-                  E-mails trackés
-                </CardTitle>
-                <Mail className="text-muted-foreground h-4 w-4" />
-              </CardHeader>
-              <CardContent>
-                <div className="text-2xl font-bold">0</div>
-                <p className="text-muted-foreground text-xs">
-                  Aucun e-mail tracké pour le moment
-                </p>
-              </CardContent>
-            </Card>
-
-            <Card>
-              <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
-                <CardTitle className="text-sm font-medium">
-                  Réponses reçues
-                </CardTitle>
-                <BarChart3 className="text-muted-foreground h-4 w-4" />
-              </CardHeader>
-              <CardContent>
-                <div className="text-2xl font-bold">0</div>
-                <p className="text-muted-foreground text-xs">
-                  Taux de réponse: 0%
-                </p>
-              </CardContent>
-            </Card>
-
-            <Card>
-              <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
-                <CardTitle className="text-sm font-medium">
-                  Follow-ups envoyés
-                </CardTitle>
-                <Users className="text-muted-foreground h-4 w-4" />
-              </CardHeader>
-              <CardContent>
-                <div className="text-2xl font-bold">0</div>
-                <p className="text-muted-foreground text-xs">
-                  Aucun follow-up programmé
-                </p>
-              </CardContent>
-            </Card>
-
-            <Card>
-              <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
-                <CardTitle className="text-sm font-medium">
-                  Boîtes mail
-                </CardTitle>
-                <Settings className="text-muted-foreground h-4 w-4" />
-              </CardHeader>
-              <CardContent>
-                <div className="text-2xl font-bold">0</div>
-                <p className="text-muted-foreground text-xs">
-                  Configurez vos boîtes mail
-                </p>
-              </CardContent>
-            </Card>
-          </div>
-
-          {/* Quick Actions */}
-          <div className="grid grid-cols-1 gap-6 lg:grid-cols-2">
-            <Card>
-              <CardHeader>
-                <CardTitle>Actions rapides</CardTitle>
-                <CardDescription>
-                  Commencez par configurer vos boîtes mail et paramètres
-                </CardDescription>
-              </CardHeader>
-              <CardContent className="space-y-4">
-                <Button className="w-full justify-start" variant="outline">
-                  <Plus className="mr-2 h-4 w-4" />
-                  Ajouter une boîte mail
-                </Button>
-                <Button className="w-full justify-start" variant="outline">
-                  <Settings className="mr-2 h-4 w-4" />
-                  Configurer les paramètres
-                </Button>
-                <Button className="w-full justify-start" variant="outline">
-                  <BarChart3 className="mr-2 h-4 w-4" />
-                  Voir les analytics
-                </Button>
-              </CardContent>
-            </Card>
-
-            <Card>
-              <CardHeader>
-                <CardTitle>État du système</CardTitle>
-                <CardDescription>
-                  Informations sur la configuration actuelle
-                </CardDescription>
-              </CardHeader>
-              <CardContent>
-                <div className="space-y-3">
-                  <div className="flex items-center justify-between">
-                    <span className="text-sm text-gray-600">
-                      Authentification
-                    </span>
-                    <Badge variant="default">Configurée</Badge>
-                  </div>
-                  <div className="flex items-center justify-between">
-                    <span className="text-sm text-gray-600">
-                      Base de données
-                    </span>
-                    <Badge variant="default">Connectée</Badge>
-                  </div>
-                  <div className="flex items-center justify-between">
-                    <span className="text-sm text-gray-600">
-                      Microsoft Graph
-                    </span>
-                    <Badge variant="outline">Non configuré</Badge>
-                  </div>
-                  <div className="flex items-center justify-between">
-                    <span className="text-sm text-gray-600">Webhooks</span>
-                    <Badge variant="outline">Non configuré</Badge>
-                  </div>
-                </div>
-              </CardContent>
-            </Card>
-          </div>
-        </div>
-      </main>
-    </div>
-  );
+  // Pass user data to client component
+  return <DashboardPageClient user={userWithRole} />;
 }
